requirements.txt
pandas>=2.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
folium>=0.12.0
plotly>=5.0.0
customtkinter>=5.0.0
geopandas>=0.12.0
numpy>=1.21.0
Pillow>=9.0.0
kaleido>=0.2.1
psycopg2-binary>=2.9.0
asyncio>=3.4.3